// workspaces/multi-source-security-viewer/plugins/multi-source-security-viewer/src/api/mssv.ts
/*
 * Copyright 2024 The Backstage Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { Entity } from '@backstage/catalog-model';
import { PipelineRunResult } from '../models/pipelineRunResult';

export type MssvApiResponse = {
  results: PipelineRunResult[];
  totalCount: number;
};

/*
 * Providers must implement the MssvApi interface
 */
export type MssvApi = {
  getPipelineSummary: (options: {
    entity: Entity;
    page: number;
    pageSize: number;
  }) => Promise<MssvApiResponse>;

  getPipelineDetail: (options: {
    entity: Entity;
    ref: string;
    page: number;
    pageSize: number;
  }) => Promise<MssvApiResponse>;
};